+++ comps/heatlist/file_based_heatlist.py
@@ -1,8 +1,8 @@
 import json
 
 from rankings.models import Couple, Dancer
-from .models import Heat, HeatEntry, HeatlistDancer, UnmatchedHeatEntry
-from .heatlist import Heatlist
+from comps.models import Heat, HeatEntry, HeatlistDancer, UnmatchedHeatEntry
+from comps.heatlist.heatlist import Heatlist
 
 
 class FileBasedHeatlist(Heatlist):
