+++ comps/heatlist/ndca_prem_heatlist.py
@@ -2,8 +2,8 @@ import requests
 import html
 
 from rankings.models import Couple, Dancer
-from .models import Heat, HeatEntry, HeatlistDancer, UnmatchedHeatEntry
-from .heatlist import Heatlist
+from comps.models import Heat, HeatEntry, HeatlistDancer, UnmatchedHeatEntry
+from comps.heatlist.heatlist import Heatlist
 
 
 class NdcaPremHeatlist(Heatlist):
